+++ src/api/semknoxBaseApi.php
@@ -212,11 +212,10 @@ use semknox\search\Struct\ProductResult;
                 CURLOPT_RETURNTRANSFER => true,
                 CURLOPT_ENCODING => "",
                 CURLOPT_MAXREDIRS => 10,
-                CURLOPT_TIMEOUT => 0,
                 CURLOPT_FOLLOWLOCATION => true,
                 CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,                
                 CURLOPT_CUSTOMREQUEST => $method,
-                CURLOPT_TIMEOUT => 30,
+                CURLOPT_TIMEOUT => 100,
                 CURLOPT_HTTPHEADER => array(
                     "Content-Type: application/json",
                     "HTTP_CLIENT_IP: ".$this->headerInfoData['clientip'],